refactor a1 fpga analysis without plotting results
